+++ SilkAndSong/HarmonyPatches/HeroController_Update.cs
@@ -8,23 +8,37 @@ namespace SilkAndSong.HarmonyPatches
         [HarmonyPostfix]
         public static void Postfix(HeroController __instance)
         {
-            if (SharedData.Level > 0)
+            if (SharedData.Level <= 0)
             {
-                float healthTime = SharedData.GetMaskSeconds(SharedData.Level);
-                if (healthTime <= SharedData.healthTimer.ElapsedMilliseconds / 1000)
-                {
-                    //SilkAndSong.instance.Log($"Health time reached: {healthTime}");
-                    __instance.AddHealth(1);
-                    SharedData.healthTimer.Restart();
-                }
+                return;
+            }
 
-                float silkTime = SharedData.GetSilkSeconds(SharedData.Level);
-                if (silkTime <= SharedData.silkTimer.ElapsedMilliseconds / 1000)
-                {
-                    //SilkAndSong.instance.Log($"Silk time reached: {silkTime}");
-                    __instance.SilkGain();
-                    SharedData.silkTimer.Restart();
-                }
+            // Make sure to pause/start based on if the game is paused
+            if (GameManager.instance.isPaused)
+            {
+                SharedData.healthTimer.Stop();
+                SharedData.silkTimer.Stop();
+            }
+            else
+            {
+                SharedData.healthTimer.Start();
+                SharedData.silkTimer.Start();
+            }
+
+            float healthTime = SharedData.GetMaskSeconds(SharedData.Level);
+            if (healthTime <= SharedData.healthTimer.ElapsedMilliseconds / 1000)
+            {
+                //SilkAndSong.instance.Log($"Health time reached: {healthTime}");
+                __instance.AddHealth(1);
+                SharedData.healthTimer.Restart();
+            }
+
+            float silkTime = SharedData.GetSilkSeconds(SharedData.Level);
+            if (silkTime <= SharedData.silkTimer.ElapsedMilliseconds / 1000)
+            {
+                //SilkAndSong.instance.Log($"Silk time reached: {silkTime}");
+                __instance.SilkGain();
+                SharedData.silkTimer.Restart();
             }
         }
     }