+++ SilkAndSong/SilkAndSong.cs
@@ -1,11 +1,12 @@
 using BepInEx;
 using HarmonyLib;
 using SilkAndSong.Helpers;
-using TeamCherry.Localization;
 
 namespace SilkAndSong;
 
 [BepInAutoPlugin(id: "io.github.danielstegink.silkandsong")]
+[BepInDependency("org.silksong-modding.i18n")]
+[BepInDependency("io.github.danielstegink.wishutil")]
 public partial class SilkAndSong : BaseUnityPlugin
 {
     internal static SilkAndSong instance;
@@ -15,13 +16,6 @@ public partial class SilkAndSong : BaseUnityPlugin
         // Put your initialization logic here
         instance = this;
 
-        // todo - modify this to install individual patches; i need to add the get language patch in the start method
-        // the language get patch needs to grab my quest description and replace it w dynamic values
-        // round bonuses to nearest 0.01
-        // for dmg bonuses, make sure to note w + or - (shouldn't be possible to get -, but i want to check regardless)
-
-        // also - go into quest manager and figure out how to add quest to manager and completion journal
-        // may not need to add it to journal, but i should check regardless
         Log($"Plugin {Name} ({Id}) has loaded!");
     }
 
@@ -29,6 +23,9 @@ public partial class SilkAndSong : BaseUnityPlugin
     {
         Harmony harmony = new Harmony(Id);
         harmony.PatchAll();
+
+        SharedData.quest = new SnsQuest();
+        WishUtil.QuestData.AddQuest(SharedData.quest);
 
         ConfigSettings.Initialize(Config);
     }